Added postgres example

+++ src/person.rs
@@ -3,5 +3,5 @@ use serde_derive::{Deserialize, Serialize};
 #[derive(Debug, Serialize, Deserialize)]
 pub struct Person {
     pub name: String,
-    pub age: u8,
+    pub age: i32,
 }
